Adds message to STDERR to show log file location
This is useful if we are trying to find out where the logs are located
and the server is running in a docker container. Looking at docker
container logs will show the location of the log file in the container,
which can lead to the mapped location on the host file system.