Updates cypher for connectivity heatmap.
The manc dataset has a number of ROIs that should not be shown in the
connectivity heatmap on the home page. These ROIs, have been marked with
an excludeFromOverview key in the meta.roiInfo data. The updated cypher
looks for this key and removes it from the list when selecting ROIs to
be used in the heatmap.