fix: ca.getDailyType_int cypher allows missing cropped attr
The CYPHER query that was selecting a cell type for the daily type api
was expecting all nodes to have a cropped attribute. If this was
missing, then no nodes were selected and nothing was returned for the
daily type. Adding an 'n.cropped IS NULL OR' condition to the CYPHER
allows it to select nodes that do not have the attribute and are
presumably not cropped.